Move DeclarativeContract to new file
